working - refactored
