+++ src/giggletech_osc.rs
@@ -4,6 +4,10 @@
 
 use async_osc::{ OscSocket, Result};
 
+// OSC Address Setup
+const TX_OSC_MOTOR_ADDRESS: &str = "/avatar/parameters/motor";
+//const TX_OSC_LED_ADDRESS_2: &str = "/avatar/parameters/led";
+
 pub(crate) fn create_socket_address(host: &str, port: &str) -> String {
     let address_parts = vec![host, port];
     address_parts.join(":")
